tinyappdev
Adding PML_REGISTER_SET_ALLOCATOR() macro to PML. This can be used in a class definition to register a method on the class which will pass the allocator used by pml_new to the object just after allocation, prior to the constructor being run. The method's with a signature should be something like 'void set_allocator(pml::Allocator *)'.
A method registered like this can't really do anything with the uninitalized object, other than assign the pointer value to a member variable, but this means that the allocator will be available inside the object's constructor when that is called (so the object can use the same allocator to allocate its members). It can also allow the object to provide a wrapper for pml_delete which uses the same allocator, so a user of the class doesn't need to remember this. The unit test test_set_allocator() shows examples of both of these kinds of usage.