tinyappdev
Build system update - should be easier to keep in sync with current 'state of the art'.
Most/all tools now have variables defined in config.mak, and are referred to indirectly e.g. '$(ECHO)' instead of 'echo' - this should make it easier to try out alternative versions of tools/compilers and so on (e.g. if we want to use a specific version of a tool which isn't on the path). We should also have some support for docs targets (using Doxygen) now, which we will test out in a subsequent checkin.